+++ src/bebone/gfx/gfx.h
@@ -12,8 +12,6 @@
 
 #include "shaders/spirv_shader_compiler.h"
 
-#include "render_target.h"
-
 #include "glfw_context.h"
 
 #endif
