+++ src/bebone/gfx/vulkan/vulkan_swap_chain.h
@@ -13,7 +13,7 @@
 #include <numeric>
 
 #include "vulkan_device.h"
-#include "../render_target.h"
+#include "vulkan_render_target.h"
 #include "vulkan_wrapper.tpp"
 
 namespace bebone::gfx {
